chore(trends): extract TrendsSeriesMeta to shared file
Address review feedback: break circular import between TrendsLineChartD3
and TrendsTooltip by moving the shared meta type into trendsSeriesMeta.ts.
